add auth permissions to user and admin

+++ routes/web.php
@@ -19,19 +19,17 @@ use Illuminate\Support\Facades\Auth;
     return view('welcome');
 }); */
 
-
-Route::get('/', [App\Http\Controllers\EventController::class, 'index'])->name('home');
-Route::get('/home', [App\Http\Controllers\EventController::class, 'index'])->name('home');
-
-/* ->middleware('auth') */
 Auth::routes();
+
+Route::get('/', [EventController::class, 'index']);
+Route::get('/home', [EventController::class, 'index'])->name('home');
 
 //CRUD Events
 
-Route::delete('/delete/{id}', [EventController::class, 'destroy'])->name('delete');
-Route::get('/edit/{id}', [EventController::class, 'edit'])->name('edit');
-Route::patch('/update/{id}', [EventController::class, 'update'])->name('update');
-Route::post('/events', [EventController::class, 'store'])->name('store');
+Route::delete('/delete/{id}', [EventController::class, 'destroy'])->name('delete')->middleware('auth','isAdmin');
+Route::get('/edit/{id}', [EventController::class, 'edit'])->name('edit')->middleware('auth','isAdmin');
+Route::patch('/update/{id}', [EventController::class, 'update'])->name('update')->middleware('auth','isAdmin');
+Route::post('/events', [EventController::class, 'store'])->name('store')->middleware('auth','isAdmin');
 Route::get('/create', [EventController::class, 'create'])->name('create')->middleware('auth','isAdmin');
-
 Route::get('/show/{id}', [EventController::class, 'show'])->name('show');
+
